interger

+++ database/migrations/2022_08_15_024013_modify_author_column_in_articles_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
     {
         DB::table('articles')->update(['author' => 1]);
         Schema::table('articles', function (Blueprint $table) {
-            $table->bigint('author')->change();
+            $table->integer('author')->change();
         });
     }
